Correção do resultado para float, para ter mais precisão

+++ lista_exec7.py
@@ -1,5 +1,5 @@
-celcius = int(input("Digite a temperatura em Celcius: "))
+celcius = float(input("Digite a temperatura em Celcius: "))
 
-fahrenheit = int(celcius*1.8+32)
+fahrenheit = float(celcius*1.8+32)
 
 print(f"A temperatura em Fahrenheit é {fahrenheit}")
